requirements.txt
flask
flask-restx
flask-httpauth
openai
pillow
werkzeug
